last version of math skills

+++ mathFunctions/StandardDeviation.go
@@ -3,5 +3,5 @@ package mathFunctions
 import "math"
 
 func StandardDeviation(numbers []float64) float64 {
-	return math.Sqrt(float64(Variance(numbers)))
+	return math.Sqrt(Variance(numbers))
 }
